requirements.txt
requests~=2.23.0
mysql-connector-python~=8.0.19
kivy.deps.sdl2
kivy.deps.glew
docutils
pygments
pypiwin32
kivy~=1.11.1


